pyproject.toml
[project]
name = "skill-manager"
version = "0.1.0"
description = "CLI tool for assembling composite skills from multiple sources"
readme = "README.md"
authors = [
    { name = "Aaron Brooks", email = "aaron.neil.brooks@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "typer>=0.12.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "httpx>=0.27.0",
]

[project.scripts]
skill-manager = "skill_manager.cli:app"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-anyio>=0.0.0",
    "respx>=0.21.0",
]

[build-system]
requires = ["uv_build>=0.9.26,<0.10.0"]
build-backend = "uv_build"
